Update ShockPot Driver.h

+++ include/ShockPot Driver.h
@@ -3,13 +3,17 @@
 #ifndef SHOCKPOT_DRIVER.H
 #define SHOCKPOT_DRIVER.H 
 
+
+#include <Arduino.h>
+
+
 const static uint8_t ShockPotPins[] = {A0, A1, A2, A3};
 
 typedef struct 
 {
-	int pin = 0;
-    int val = 0;
-    int pos = 0;
+	uint8_t  pin = 0;
+    uint16_t val = 0;
+    uint16_t pos = 0;
 } ShockPot;
 
 
@@ -23,6 +27,7 @@ public:
     int  getReading();
     void getPosition();
 };
+
 
 static ShockPotDriver ShockPot_;
